Fixes for arm64 builds
Some minor fixups for building natively on arm64, including using the
correct sources for Go and Rust, and setting the target ARCH for the
kernel build.

Then some updates to GRUB and rpm2img to support an EFI partition on
arm64. This creates an undersized EFI partition if building on arm64
which luckily fits the GRUB binary and means the partition sizes between
x86_64 and arm64 match.

There are still other package-specific issues that prevent
cross-compiling which will have to be fixed in a later patch.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>