buildsys: fix incremental builds alongside fetches
Fixes #484

Using a marker file in the Cargo specified $OUT_DIR we are able to run
through the extent of fetches by using the same environment variable as
before and printing of an intentionally stale [^1] file (which doesn't
exist at the fetch time). This stale file causes Cargo to rerun the
build.rs script on the next call to it and during that run - in
environment which the variable `BUILDSYS_BUILD_FETCH_ONLY` doesn't exist
- the marker file is created to reinstate incremental build support by
way of Cargo's rerun-if* directives.

This change retains the property of rebuilding of sources on other
changes to the source files in the Cargo package and also by using the
$OUT_DIR to record this progress we also extend support for
per-"context" incremental builds as defined by Cargo with respect to
compilation and other configuration sets [^2].

[1]: https://github.com/rust-lang/cargo/blob/ba5cca20acc6b61eec24e35b3fd146fa55899b6f/src/cargo/core/compiler/fingerprint.rs#L401-L404

[2]: https://github.com/rust-lang/cargo/blob/759431fe98c5e21546ef587945eb0a78cb60f331/src/cargo/core/compiler/fingerprint.rs#L34-L62

Signed-off-by: Jacob Vallejo <jakeev@amazon.com>