build: remove cascade feature
Previously, doing a full build that respected the dependency graph
would cause almost every package to be rebuilt when the kernel was
updated, which happened at regular intervals.

The cascade feature was added to allow a developer to opt out of the
full dependency tracking in the interests of speeding up iterative
builds. However, making use of it required two full builds to get
cargo in sync, with no reliable way to predict whether either build
would be quick. This limited its usefulness in practice.

Now that the kernel is a leaf package, we will only rebuild the world
when glibc changes, which we expect to happen less often.

Signed-off-by: Ben Cressey <bcressey@amazon.com>