Allow for a separate API model per variant
Pull #578 added support for build variants to support different features and
use cases, starting with aws-k8s and aws-dev.  These variants use different
software, and therefore need different configuration.  This change allows for
having a different API model (and default values) per variant.

The model was moved to workspaces/models.  This crate actually contains the
model for each variant in separate subdirectories.  Cargo's `build.rs` is used
to symlink in the proper source code before build.

The "proper" source code is chosen using the VARIANT environment variable.  The
build system passes this through using the value of BUILDSYS_VARIANT, which is
set in the top-level Makefile.toml and can be overridden on the command-line.
For local cargo builds, the developer must set the VARIANT environment
variable; we don't currently want to choose a default model.

To start out, the aws-k8s and aws-dev models (and related default values) are
copies, with aws-dev having the Kubernetes-related values removed.  The next
step is to figure out an appropriate method for sharing structure; until then,
we'll need to duplicate most changes.