Check licenses before building rather than after
If a license is incorrect and will be rejected, it's good to know that before
you've spent time building the project.

As an example, if the license in a Cargo.toml file is wrong and will be
rejected, but we check after the build, you have to rebuild that package even
though none of the source has changed.

This can be overridden by building with
BUILDSYS_ALLOW_FAILED_LICENSE_CHECK=true in case you're doing local development
and don't have license information yet.