updata: allow specifying a start time for waves
Previously, wave start times would always be relative to when you ran updata,
meaning you'd have to hold off on running updata until you were ready to start
a release process.  This lets you (optionally) pick any start time for the
start of the waves, and the given start_after offsets will be relative to that
time.

To keep the code and the deployment process simple, this also requires
specifying relative offsets in wave definition files, rather than absolute
times; we only used relative offsets anyway so that the files were reusable and
it was easier to understand the delays between waves, but this codifies that
behavior.