updog: remove jitters, calculate wave position to determine availability
This removes jittering in updog and instead use the seed to find a
target time position within the wave to determine whether an update is
available.

This also removes the --timestamp option since it's no longer necessary
without updog returning the jittered time.

As a result of this change, updates will only be visible if they are
available (host is past their defined update wave). To see all updates
regardless of update waves, either --ignore-waves needs to be specified
or settings.updates.ignore-waves needs to be set to true.