pubsys: change "parameter overrides" to "conditional parameters"
Rather than template directories with override files for specific variants and
architectures, this uses a single TOML file for all parameters with (optional)
keys restricting parameters to specific variants and arches.

This model is clearer, based on early feedback, and also allows users to
specify parameters that are conditional to a variant *and* an architecture,
whereas the override files were more like "or" matches.