Makefile: remove `tasks.publish-setup-without-key`
For all tasks that depend on both `pubsys-setup-without-key` and `pubsys-tools`,
`tasks.fetch-sources` is invoked twice when it's not suppose to.

Apparently cargo-make does not check if a task has already ran if
you use the `run_task` directive.

This adds some unnecessary overhead to pubsys repo validation tasks (+1 seconds
per invocation).

By setting `ALLOW_MISSING_KEY` env var in the tasks we need to `pubsys-setup`
without a signing key, we avoid needing a separate `tasks.pubsys-setup-without-key`.