pubsys: resolve rustls in coldsnap and tough
Update coldsnap to v0.2.0 and set it to use rustls. This is a necessary
prerequisite of any tough-related updates. When we update tough, then
coldsnap and rough will both resolve to rusoto 0.45 and their feature
selections will conflict. By bumping coldsnap and setting it to rustls,
we are now able to update tough, tough-kms, or tough-ssm.

This commit also explicitly sets the rusoto crates to use rustls. This
was also found to be necessary when newer versions of tough* are
introduced.