buildsys: use argument vecs instead of string manipulation
Some arguments have internal spaces, like the list of packages in --build-arg
PACKAGES=a b c.  This was handled using a "|" marker that was later replaced by
spaces.  For other uses of --build-arg, like PRETTY_NAME, we don't want to
arbitrarily restrict the characters that can be used, in case someone wants "|"
in their OS name.

This change stores arguments directly in vectors, rather than strings that we
split on space, so that we can pass through arbitrary characters.