buildsys: Add ability for variants to specify supported architectures
This change adds an additional key that can be specified in a variant's
`Cargo.toml`, `supported-arches`.  It is a list and the supported enum
values are `x86_64` and `aarch64`.  If `supported-arches` is specified,
the current `BUILDSYS_ARCH` is checked against the list.  If
`supported-arches` is not specified, the build continues as before.