Makefile: Add `build-ova` target
This change adds a new `build-ova` target that will build OVA packages
for variants that build VMDK images.  The OVA packages are named
identically to our images and end up in a new directory `build/ovas`. To
support the building of OVA packages, a new OVF template is included in
`variants/vmware-dev/template.ovf`.  When building an OVA package, the
current values are substituted into the template and it is included in
the final OVA package.