refactor kmod kit creation
Now that the installed kernel is a property of the variant, we need
to ensure that the kmod kit includes the development files that match
the chosen kernel.

By creating it as another stage in the variant build, it's easier to
find the right files, and to guarantee that the kmod kit is recreated
whenever the image changes.

Signed-off-by: Ben Cressey <bcressey@amazon.com>