build: use a friendly name for OVAs
This makes OVA building a mandatory step and stores the output with
other variant artifacts.

Convenience symlinks are created in "latest", along with a symlink
that's used to give the repo artifact a friendly, predictable name.

Since the archives directory is no longer in use, remove it from the
setup and cleanup steps.

Signed-off-by: Ben Cressey <bcressey@amazon.com>