Makefile: make OVF file first in the OVA bundle
This change moves the OVF file to the beginning of the OVA tar bundle to
align with the OVF spec.  Some client software is nice enough to take
care of this for you and upload the OVF first.  vSphere however, will
choke when a user provides a URL to an OVA bundle where the OVF file
isn't first.