Makefile: Add the ability to override the AMI data file suffix
This adds an additional variable to the `Makefile` that can be used to
override the suffix of the file where AMI data is written during publish
operations.  It is currently set to `amis.json`, which is identical to
its previous hard-coded value.

Overriding this variable can be useful when a user is publishing to
multiple places with different `Infra.toml` files and wishes to write
AMI data to specifically named files.