Makefile: Add disk image files to TUF repo during `cargo make repo`
This change adds the disk image files, if they exist, to the targets
that are linked into the TUF repository during the `cargo make repo`
target.  When the image files are in the repo, we can download them with
`tuftool` which can simplify the process of creating AMIs at a different
time or from a different machine.