rpm2img: Conditionally add bootconfig parameters to grub menuentry
This change conditionally adds the `bootconfig` parameter to the kernel
command line if `set-private-var` is added to a variants
`grub-features`.  The `bootconfig` parameter signals the kernel to look
for the Boot Configuration at the end of the specified initrd.  This
change also adds the corresponding initrd entry to the grub menuentry
which points at the bootconfig file on the `BOTTLEROCKET-PRIVATE`
partition.  Bottlerocket technically doesn't use an initrd, so this
initrd will only contain Boot Configuration data.  If the file doesn't
contain Boot Config data, the kernel logs a message but does not fail to
boot.

An empty `bootconfig.data` file is created and added to the
`BOTTLEROCKET_PRIVATE` partition to keep grub from pausing and printing
an error that the file doesn't exist.