buildsys: Add `grub-features` property to variant config
This change adds a `grub-features` property to the supported configuration
that can be set in a variant's `Cargo.toml`.  This property is meant to gate
certain features of grub and it's config to variants that support it.  Since
grub isn't updated during OS update, it's important to ensure that variants
which are known to support a certain grub feature receive the proper grub
config.

Currently, the only supported parameter is `set-private-var`.  This
parameter means the variant has a grub config that understands how to search
for and find the BOTTLEROCKET_PRIVATE partition, and set the corresponding
`$private` variable.