Use generated network configuration by default
This commit removes the hard-coded `eth0.xml` network configuration file
in favor of using `netdog` to generate the network configuration for all
variants.  It adds a new systemd unit file to run `netdog
generate-net-config` early in boot, before the network is up.  To
generate the network configuration for AWS/VMware variants, we pass
`netdog.default-interface=eth0:dhcp4,dhcp6?` which `netdog` interprets
and generates network config similar to the hardcoded file previously
used for these variants.

For metal variants, we decided to use systemd-udevd's predictable device
naming so we can count on network devices being named identically every
boot.  We currently pass `net.ifnames=0` on the kernel command line
which disables predictable naming, which is fine for AWS and VMware
variants as hardware is controlled and instances typically initially
come up with a single interface in the same PCIe location.  In order to
continue using `net.ifnames=0` for AWS/VMware, we move this parameter
out of the default kernel command line to the KERNEL_PARAMETERS section
of each variants `Cargo.toml`.

Aside: We previously passed `biosdevname=0` on the kernel command line
as well.  `biosdevname` is a udev helper utility written by Dell for
consistent device naming based on SMBIOS info.  We don't currently
package the helper or include the udev rule that uses it, so we have
removed the `biosdevname` parameter entirely.