Makefile: Add targets to create and validate Boot Configuration
This change adds two new targets to the Makefile.  Both make use of the
`bootconfig` tool in the latest version of the SDK.  The first new
target, `boot-config`, gives users the ability to create a properly
formatted Boot Configuration initrd given a valid config file.  The
second new target `validate-boot-config`, validates the Boot
Configuration initrd, listing its contents.  If the initrd is somehow in
a bad format, the tool will fail and print an error.

The Makefile targets expect a valid configuration file in the root of
the Bottlerocket repo named "bootconfig-input".  The Boot Configuration
initrd will be created in the root of the repo with the name
"bootconfig.data", which is what Bottlerocket expects the file to be
named when using it.