variants: use quiet parameter for aws and vmware
Use the "quiet" parameter for platforms where the hardware is known
and the kernel output is not generally needed on the console in order
to diagnose issues. This includes the "aws" and "vmware" variants.

The "metal" variants continue to use the more verbose default output
so that any hardware without matching driver support can be detected.

Force systemd to always print the status of units, since this can be
invaluable data when troubleshooting failures caused by dependency
ordering problems.

Remove the printk settings from the default sysctls, since these will
be applied by `systemd` very early and override the "quiet" behavior.
The default values in `/proc/sys/kernel/printk` are controlled by the
kernel config, and will be "7 4 1 7" without the "quiet" parameter,
and "4 4 1 7" when it is present.

The main motivation for this is to improve boot performance and make
it more consistent. printk logging is expensive, and can slow down
many code paths in the kernel.

Signed-off-by: Ben Cressey <bcressey@amazon.com>