build: avoid setting default build flags
Fedora 36 automatically sets certain flags during RPM builds:
  https://fedoraproject.org/wiki/Changes/SetBuildFlagsBuildCheck

Since the `set_cross_build_flags` macro will use these flags if they
are specified in the environment, this causes the expected flags for
cross-compilation to be ignored, leading to build failures.

It's necessary to undefine the macro to suppress the new behavior.
Unfortunately, that can't be done through the usual macro inheritance
approach, because macros cannot be undefined, only reset. Therefore
the `rpmbuild` invocation is modified to remove the definition.

Signed-off-by: Ben Cressey <bcressey@amazon.com>