build: move RPM database back to /var
Fedora 36 moved the RPM database to /usr:
  https://fedoraproject.org/wiki/Changes/RelocateRPMToUsr

For Bottlerocket, /usr should be a symlink to the sysroot directory.
If the RPM database is stored in /usr, then it will be created prior
to installing any Bottlerocket packages, and create a conflict with
the intended symlink.

Ordinarily this would be changed in the shared macros, but doing that
could confuse `rpm` during the package build step, since the SDK's
rpm database is still in `/usr` and package builds require a mix of
host and target packages to be installed.

Instead the definition is overridden just before image creation, when
host packages are no longer needed and the target packages have not
yet been installed.

Signed-off-by: Ben Cressey <bcressey@amazon.com>