buildsys: generate OVAs during image builds
The larger goal of this change is to make image builds more hermetic,
rather than relying on additional scripts to run afterwards to create
files and symlinks. That allows the artifacts in the directory to be
archived and extracted elsewhere more easily.

The main historical blocker for generating OVAs along with VMDKs was
that the published image sizes were specified through environment
variables rather than in the variant manifest. Now that the manifest
provides these values, they can be passed through to image builds.

The bulk of this change consists of moving the OVA logic around, and
dropping the unnecessary size checks.

As part of this change, some OVF template fields are also renamed to
standardize on the "os volume" designation for the primary full disk
image. Previously this was called the "root volume" in places, which
can lead to ambiguity because there is also a "root image" used for
in-place updates.

Signed-off-by: Ben Cressey <bcressey@amazon.com>