pubsys: use variant manifest for AMI volume sizes
Switch pubsys over to use the variant manifest to determine published
AMI volume sizes, rather than relying on environment variables.

This aligns AMIs and OVAs around the variant manifest as the source
of truth for properties related to publication, which is a necessary
step for Secure Boot support.

As part of this migration, some fields and variables are renamed to
standardize on the "os volume" designation for the primary full disk
image. Previously this was called the "root volume" in places, which
can lead to ambiguity because there is also a "root image" used for
in-place updates.

Signed-off-by: Ben Cressey <bcressey@amazon.com>