build: refactor grub features as image features
Presently there is only one image feature flag, indicating whether
GRUB has the `search` module built-in along with a config file that
sets the "private" variable.

For Secure Boot, we will end up with additional image-level feature
flags. One will be for GRUB, to indicate whether to enforce a signed
config file, and the other will be to enable UEFI Secure Boot.

Since not all of these flags will necessarily be tied to GRUB, take
this opportunity to migrate to a generic "image features" structure
to track the flags before any more are added.

Signed-off-by: Ben Cressey <bcressey@amazon.com>