buildsys: fix package rebuilds for feature flags
The previous approach for detecting changes to image feature flags
that should trigger a rebuild was incorrect. It would cause packages
to be rebuilt whenever the variant manifest that was the target at
the time of last rebuild changed.

This would happen regardless of whether image feature flags changed,
whether the same flags were set for the current variant manifest, or
whether the flags were even used by the package.

Fixing this requires two prerequisites.

First, the current image feature flags must be set in the environment
prior to invoking `cargo build`, or it will be too late to emit the
`cargo:rerun-if-env-changed` directives, since the environment that
`cargo` runs in is not influenced by changes from the build script.

Second, each package must declare which image feature flags it cares
about, if any, so that the correct cargo directive can be omitted.
The environment does not carry this information, since the absence of
a particular environment variable can also mean that a package needs
to be rebuilt to disable a previously-enabled feature.

With this functionality in place, each package build can emit the
directives needed for correct change detection. For robustness, only
the feature flags tracked by the package are passed through to the
build, meaning the feature will always be treated as "disabled" by
bconds in the spec file.

Fixes: eb44f13a7877 ("build: pass image features to package builds")

Signed-off-by: Ben Cressey <bcressey@amazon.com>