sources: fix clippy warnings
Adopt the fixes suggested by clippy in most cases.

Disable pedantic warnings for the update metadata crate after fixing
several. The intent seems to have been to hold the updater code to an
especially high bar, but in practice it creates a higher maintenance
burden when updating Rust, and requires changes in mission-critical
code for merely stylistic reasons.

Signed-off-by: Ben Cressey <bcressey@amazon.com>