Update docker run `--security-opt` command format
During our build process we have several places where we use the
Bottlerocket SDK to perform the build. In these `docker run` commands we
pass `--security-opt label:disable` to avoid labeling a large number of
files.

It appears this syntax may have changed since we started using it, and
the correct syntax is no `label=disable`. Based on reports in the Docker
repo, the way we had it of `label:disable` may not actually work.

While there are other issues to be addressed, the older format generates
an error when trying to build with finch, while the newer format -
though not recognized - only generates a warning message.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>