rpm2img,release: conditionally create data fs based on part layout
DATA-A partition in the "split" layout is too small to create an
initial filesystem with desirable filesystem parameters. So we defer the
filesystem creation to later at runtime with 'prepare-local-fs' after
the partition gets resized to fill the disk/volume. This only needs to
happen when DATA-B on the separate disk/volume is unavailable and we
have additional space on the root disk/volume for DATA-A to grow.

For "unified" layout, DATA-B does not exist since we only anticipate the
root disk/volume and no other storage device. Therefore we can skip
creating a filesystem for DATA-B.