buildsys: Add user agent header in http request
Buildsys does not set the user agent in HTTP request when fetching
upstream sources. But some servers may not like the requests without a
user agent header and netfilter.org is one of them. All these
requests succeed when a user agent is set, so include the header in
requests that buildsys sends out.