docker-go: change docker run mounts
Twoliter will be calling this script from within a container that has
the host's docker socket mounted. It is necessary to match internal
and external paths for any state that we wish to preserve on the host.

Because the data we want to cache is at $GOPATH/pkg/go, it was necessary
to create this as a subdirectory under .gomodcache and mount that so
that we could point GOPATH to .gomodcache.