simplify project file
At the moment, when we are only using Twoliter as a wrapper for
cargo make, the project file does nothing other than establish the
location of the build root. For now, instead of causing Bottlerocket to
check in confusing data in its Twoliter.toml file, let's simplify remove
all keys from the schema.