pipesys: add tool to share file descriptors
For Bottlerocket, package and image builds both involve copying large
binary files into and out of builds. When using Docker to create the
build environment, all this data must pass through the Docker daemon,
which balloons its resident memory usage in unpredictable ways and
can lead to crashes or OOM kills.

Docker also implements caching for intermediate layers. While useful
in the general case, this is unhelpful for these builds, because a
minor edit for troubleshooting purposes can produce gigabytes of data
that will never be useful again but that remain in Docker's cache.

Docker understandably does not provide a straightforward way to bind-
mount a host directory into a build, as that would decouple the build
from local inputs that are potentially required for it to succeed,
making it difficult for others to reproduce the build. Since buildsys
manages those inputs for Bottlerocket builds, this is not as much of
a concern, but the limitation remains.

On Linux, Unix domain sockets can be used to pass file descriptors.
They can also be abstract, with no backing socket file, and only
accessible within the network namespace of the opening process. This
can be used to bend Docker's rules and provide the equivalent of a
bind mount, with enough effort.

pipesys provides an abstraction over these capabilities, in the form
of library code to serve and fetch file descriptors. It can run as a
standalone program to serve a path's file descriptor to any clients
that send a request. It can also work as a bind mount substitute, by
retrieving a file descriptor and creating a symlink to it for easy
access from subsequent programs.

Signed-off-by: Ben Cressey <bcressey@amazon.com>