twoliter: switch to fixed directory for tools
Secrets have a maximum size of 500 KiB, which is not sufficient to
pass binaries to the build. This would prevent us from using native
code rather than shell scripts.

Install the tools into `build/tools` instead, taking care to clean up
any previous installs and to reset the file modification times so as
not to bust any caches or trigger unnecessary rebuilds.

Signed-off-by: Ben Cressey <bcressey@amazon.com>