allow cmd exec quiet
Make it possible in Twoliter to exec an external command quietly.
Previously this looked at the log level filter to determine whether or
not to send the command output to stdout/stderr. Now there is a form
of the exec command that allows you to specify whether or not to send
the output to stdout/stderr.