buildsys: use println instead of eprintln
Unless the process fails, Cargo will not print stderr. This can be
a difficult when debugging an issue that led to a downstream process to
fail. This commit converts a couple of eprintln statements to println
because these statements are not in guaranteed-to-fail code paths and
will not be seen if the program exits cleanly.