buildsys: fix lookaside cache error swallowing bug
If an error occurred while fetching from the lookaside cache, and
we were told not to fallback to upstream sources, then the fetch error
would be swallowed instead of propagated. This might lead to a harder
to understand downstream error when the tarball was missing.

Now we propagate the fetch error.