buildsys: fix image features
This commt fixes a couple of regressions introduced in 92bc3bb. Most
notably, the wrong set of image_features was being used to set package
build args. In this commit we correctly use the image features found
in the downstream variants Cargo.toml.

Also fixes an oversight where an error constructing the Builder object
would have resulted in a panic.