twoliter: generate metadata rpm for variant build
Create an RPM to hold metadata about the variant we're about to build
and add it to the set of packages that gets installed.

The virtual "Provides:" in the metadata package will allow other RPMs
to express forward and reverse dependencies on variant properties,
such as the platform or the enabled image features.

This can be useful in cases like `early-boot-config`, where certain
user-data providers should always be installed when building an "aws"
variant. It's also expected to help for the FIPS image feature, where
many packages will have FIPS and non-FIPS subpackages, and listing
them all in the variant definition would be tedious and error-prone.

Signed-off-by: Ben Cressey <bcressey@amazon.com>