twoliter: adapt to the unified SDK
Since the SDK contains toolchains for both supported architectures,
we no longer need either a separate toolchain image or the logic to
automatically suffix the image name with the target architecture.

Make a breaking change to the `Twoliter.toml` schema to reflect that
only a single SDK image is needed. This image can be specified with
registry, repository, and tag components rather than using terms that
don't map to the OCI distribution spec. The increase in clarity and
simplicity is judged to be worth the one-time migration cost from the
compatibility break.

Since either architecture is supported by the SDK, we no longer need
to pass any target architecture value through the code. The host's
architecture still matters, in the sense that the SDK must either be
an image that runs on that architecture or a manifest that includes
such an image. That was an implicit requirement before and remains
implicit.

Signed-off-by: Ben Cressey <bcressey@amazon.com>