rpm2img: add support for bootconfig drop-ins
This allows bootconfig data to be installed to `/boot/boot-config.d`
by packages, where it will be combined into the default boot config
for the image.

Packages can unconditionally install snippets if they should be used
for any image with that package installed. They can also make use of
RPM strong and weak dependencies to arrange for the snippet to only
by present if a certain image feature is enabled.

Signed-off-by: Ben Cressey <bcressey@amazon.com>