unplug: add tool to restrict network access
For pipesys to provide file descriptor accesses via an abstract UNIX
domain socket into container builds, the build must run with in the
host network namespace.

That conflicts with the requirement that packages are built with no
access to the network. This ensures that builds cannot have unmodeled
dependencies on remote resources that could cause them to break or
produce different results across executions.

To satisfy the requirement, create an "unplug" tool that blocks most
socket-related calls. Existing solutions like firejail or nsjail can
also do this, but they expect to run in a privileged context in order
to set up namespaces and the like, and build containers do not have
these privileges.

Signed-off-by: Ben Cressey <bcressey@amazon.com>