rpm2img: regenerate kernel module info if possible
With conditional compilation being eliminated, it is no longer always
the case that all of the modules built by the kernel package will be
installed on all variants. For example, some network device drivers
may be built as modules and only installed on bare metal variants, if
the corresponding hardware is not expected to show up in virtualized
environments.

This requires some care in packaging so that any module dependencies
that are not provided by the kernel will be provided by other modules
that are also installed. By regenerating module info at image build
time, it's guaranteed to be present and correct, or otherwise the
build will fail.

Signed-off-by: Ben Cressey <bcressey@amazon.com>