makefile.toml: one workspace and cargo metadata
For kit builds, Buildsys will need to understand the cargo dependency
graph. This requires a call to cargo metadata which, for efficiency
sake, we do only once before calling cargo build. This commit updates
Makefile.toml to provide that information to Buildsys.

Additionally, we want to have one workspace at the top-level of the
Twoliter project which includes all packages, kits and variants.
This commit also includes changes to the Makefile.toml to support a
single workspace.