twoliter: error if old rpm structure is discovered
Add a check to Makefile.toml to see if the user has rpms in the
build/rpms directory. If so, this is likely a problem because the new
version of Buildsys will expect per-package subdirectories, and Cargo
will not know it needs to rebuild.

Raise an error that informs the user they need to run cargo make clean.