buildsys: deprecate build-package variant sensitivity
Packages can no longer depend on knowing what variant they are being
built for, so we raise an error if they declare variant-sensitivity or
package-features in their Cargo.toml.

Bottlerocket is not quite ready for this, though. So we place this
change behind a feature flag variable:
BUILDSYS_DEPRECATED_FEATURE_VARIANT_SENSITIVITY