buildsys: include commit timestamp in package Release, set Epoch
Before the onset of kits, Bottlerocket would build RPMs with Release
values set to the commit of Bottlerocket being released (or, more
accurately, the commit of Bottlerocket being built). This worked fine
under the monorepo model, but with the advent of kits, this is not
enough for indicating incremental changes to a project. There must be
some monotonically increasing value present in the application-invetory
for a Bottlerocket variant for systems to key off of for versioning and
Release comparisons.

This change computes the unix timestamp in ms of the project commit from
which a package is built by buildsys and sets the Release of the package
as <timestamp>+<commit short sha>. It also sets the Epoch value of RPMs
and their subpackages based on an environment variable
BUILDSYS_VERSION_BUILD_EPOCH.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>