twoliter: skip missing per-package dirs
This fixes backwards compatibility when moving between releases when
the twoliter version changes but some of the packages (such as glibc)
do not. The cargo metadata for the older builds is still valid, and
the RPMs will be present in the top-level output directory, but the
per-package subdirectories will not yet exist.

Signed-off-by: Ben Cressey <bcressey@amazon.com>