build: no longer ignore file type for actions
GitHub actions runs are not taking place when they need them to because
one of the ignored file types was changed. This is not the desired
behavior. We would want the actions run to take place if any file not
matched by "ignore file" was touched, and this is not happening. So,
stop using the feature.