twoliter: combine build and repack Dockerfiles
The build and repack Dockerfile were separate because repack needs
access to the disk and filesystem image files that were too costly
to copy through the build context for all builds.

Now that bulk binary data is passed through pipesys, the stages for
repack can be merged back into the primary Dockerfile.

Signed-off-by: Ben Cressey <bcressey@amazon.com>