twoliter: create per-package layers in rpm2kit
Instead of combining all packages and repodata into a single layer,
which will be different regardless of how many of the packages were
changed, add one layer per package and a final layer for repodata.
That way, many layers can be pushed in parallel, and layers can be
skipped if they are already present in the remote repository.

Signed-off-by: Ben Cressey <bcressey@amazon.com>