revert application-inventory: use RPM's Version and Release, set Epoch
revert most of #263, which set the Epoch for packages built with
Twoliter, set package inventory Version as the actual package
version, and built a customer release field to give to the package.

This PR retains the change to setting Release in the actual package RPM
but reverts the change to include that in the application inventory for
the package.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>