rpm2kit: find first package via awk
`head` will exit as soon as it reads the desired number of lines, and
potentially without reading all data from the stdin pipe. That leads
to occasional SIGPIPE errors which cause the build to fail.

Instead, tell `awk` to read all input and print the first line before
it exits.

Signed-off-by: Ben Cressey <bcressey@amazon.com>