application-inventory: use core-kit version for package versions
In order for package version comparisons to be valid in the OOTB and
kits style builds of Bottlerocket, application inventory is generated to
list the version of a package by where it was sourced from. This will be
the kit version.

This commit is a first iteration on this approach and must be extended
in future work to apply to any external kit and not just the
bottlerocket-core-kit, although that is the only external kit for now.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>