twoliter: refactor build.Dockerfile.dockerignore
When using pipesys for inputs and outputs, most of the build context
can be excluded. The exceptions are source files that are useful to
snapshot at the moment the build is initiated, such as package specs
and first-party code under sources, along with tools such as pipesys
itself that must be present in the build.

Current versions of `cargo` do not work with a completely read-only
copy of the crates in `CARGO_HOME`, so that directory must be sent
via build context even though it is nominally read-only.

Signed-off-by: Ben Cressey <bcressey@amazon.com>