buildsys: add more image feature flags
Promote "in-place-updates" and "host-containers" to feature flags,
and enable them by default for backwards compatibility.

The in-place update feature requires two sets of partitions, which
wastes disk space if the variant will only ever run on nodes that are
replaced rather than updated in-place.

The host containers feature supports execution of containers at fixed
points during the host's lifecycle: during boot, and as background
services once the host is fully configured. This may not be desirable
for variants looking for stronger guarantees that only the software
included in the immutable root filesystem will ever run.

Expose these as feature flags so that variants can opt-out.

Signed-off-by: Ben Cressey <bcressey@amazon.com>