twoliter: support new image feature flags
Add the "in-place-updates" and "host-containers" features to the
metadata spec so that packages can arrange to be installed based on
whether they are enabled or disabled.

Disabling in-place update support implies removing the second set of
partitions that will never be used. Adjust the partition sizing logic
to double the size of the first set of partitions when the second set
is omitted.

Overhaul the "get_partition_sizes" logic in partyplanner to avoid
making assumptions about the partition layout of input images, such
as how many there will be, and instead to just record details about
each partition.

Signed-off-by: Ben Cressey <bcressey@amazon.com>