imghelper: hoist AWS vars into global environment
When secure boot signing was refactored in 310cadd8, the code that
exports AWS credential variables into the global environment was moved
into a bash function.

When used in a bash function, the `declare` keyword makes the declared
symbol local, unless the -g option is also supplied. This adds the -g
option to export the variables into the caller's environment as
intended.