buildsys: add CI/CD hack
For local builds, the detection logic around when to rebuild crates
works fairly well outside of edge cases like upgrading Rust.

However, the same is not true for reusing cached builds with CI/CD.
`cargo` relies on file modification times, and Git does not preserve
these across checkouts, making it difficult to get unmodified files
back to the state that `cargo` expects in order to avoid a rebuild.
In addition, the fingerprint that `cargo` assigns for a given crate
is not consistent across checkouts or changes to the Rust toolchain,
which again leads to unnecessary rebuilds.

To sidestep these challenges, add a `BUILDSYS_CICD_HACK` environment
variable that causes `buildsys` to exit before launching any Docker
builds, and after downloading external files and emitting the change
detection commands. This assumes that the build output directory is
complete and current, which is not possible to prove without letting
the build run. If that assumption holds, then the result is that the
crates produced by `cargo` will be regenerated and will be primed to
track subsequent changes.

Signed-off-by: Ben Cressey <bcressey@amazon.com>