buildsys: drop bottlerocket-variant binary
Previously, this functionality was in a separate binary because the
environment variables needed to be set prior to `buildsys` invocation
so that `cargo` could track them for change detection purposes. Now
that packages can't be conditionally compiled based on the variant,
change detection isn't required, so the binary can be removed.

Signed-off-by: Ben Cressey <bcressey@amazon.com>