buildsys: drop variant arg for variant subcommands
The switch from relying on BUILDSYS_VARIANT to CARGO_MANIFEST_DIR did
not take into account the case where `buildsys` is not executed by
`cargo`, and hence CARGO_MANIFEST_DIR is not set in the environment.

This broke the repack-variant subcommand, which receives the value
from a command-line argument when invoked through Makefile.toml.

In any case, having both `variant` and `cargo-manifest-dir` args that
are controlled by the same environment variable is redundant. Fix the
bug and remove the confusion by using `cargo-manifest-dir` alone to
identify the artifact that `buildsys` is meant to act on.

Fixes: 4526d6ee ("buildsys: stop relying on BUILDSYS_VARIANT")

Signed-off-by: Ben Cressey <bcressey@amazon.com>