buildsys: mark erofs feature as experimental
erofs is a relatively new filesystem, especially when compared with
ext4. It shows promising results in terms of improved boot times, but
its behavior at runtime on fully loaded nodes is more difficult to
predict.

Mark the image feature as experimental for now. This allows it to be
enabled by downstream variants to facilitate additional experiments,
without immediately endorsing it for production use.

Signed-off-by: Ben Cressey <bcressey@amazon.com>