twoliter: avoid duplicate macro declaration
Processes that run as root need the target's macros in order to parse
spec files. Copy the macros to root's home directory rather than to
the system-wide `/etc/rpm` location, to avoid defining macros twice
for other users, such as the unprivileged builder user. Otherwise,
the redundant definition forces macros to be undefined twice to clear
them from the per-user and the system-level scopes.

Signed-off-by: Ben Cressey <bcressey@amazon.com>