twoliter: ensure krane tmpfile is deleted
When using named tempfiles or tempdirs, the `tempfile` crate relies on
Rust calling `drop()` in order to delete the created temporary files.

We were using a static reference to a tempfile of the krane binary,
which is not guaranteed to drop when the program completes. This caused
twoliter to leak somewhat large krane binary tempfiles.

This commit moves away from a static reference to the tempfile, instead
improving ergonomics around ImageTool owning the tempfile reference.
This allows the Rust compiler to more-reliably drop() the reference on
exit.