test: add integration test for application-inventory generation
Introduces an integration test that will build bottlerocket twice.
First it will build bottlerocket with the latest released version of twoliter.
Then it will build bottlerocket with the current commits build. Afterwards
it does a comparison to make sure the application-inventory.json file is generated
correctly and matches the released version.