rpm2img: reclaim reserved space on ext4
ext4 reserves 5% of the filesystem for the superuser to avoid situations
where non-privileged processes or users use up the disk space and the
superuser needs some space to clear the usage. Since the root filesystem
we are creating in rpm2img is read only, this reserved space is not
needed. Disabling this reserved space with -m 0 results in disk
utilization showing the same number of blocks in use, but available is
5% higher.

The ext4 feature for online resizing reserves additional blocks for
making it possible extend the block group descriptor table when making a
filesystem larger. Since we will never resize this filesystem, it is
safe to disable it with ^resize_inode and the filesystem provides
additional blocks for usage. This frees up about 2% of the blocks.

The root filesystem doesn't need backup super blocks and block
descriptors so using the stricter sparse_super2 saves a bit on space.
Since the file system is typically small, the savings is small as well.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>