rpm2img: Emit artifact-metadata.json with disk utilization
Add helper functions and invocations to rpm2img to get the utilization
of a partition and emit this into a file called artifact-metadata.json
with data for ROOT-A and BOOT-A partitions. This can be extended to
provide other metadata about the images but this helps track the
resulting utilization of the various partitions created via an image
build. The change is mirrored in img2img so both scripts produce the
same file.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>